Separating PhysicsEntity from Entity
