+++ core/src/com/superduckinvaders/game/entity/Entity.java
@@ -11,16 +11,6 @@ import com.badlogic.gdx.math.Vector2;
  * Represents an object in the game.
  */
 public abstract class Entity {
-    public static final short WORLD_BITS = 0x1;
-    public static final short PLAYER_BITS   = 0x2;
-    public static final short MOB_BITS   = 0x4;
-    public static final short PROJECTILE_BITS   = 0x8;
-    public static final short ALL_BITS   = WORLD_BITS | PLAYER_BITS | MOB_BITS | PROJECTILE_BITS;
-    public static final short NO_GROUP  = 0;
-    public static final short MOB_GROUP  = -1;
-    
-    public short categoryBits = PLAYER_BITS;
-
     /**
      * The round that this Entity is in.
      */
@@ -65,51 +55,13 @@ public abstract class Entity {
         this.x = x;
         this.y = y;
     }
-    
-    public void createStaticBody(short categoryBits, short maskBits, short groupIndex, boolean isSensor){
-        createBody(BodyDef.BodyType.StaticBody, categoryBits, maskBits, groupIndex, isSensor);
-    }
-    public void createDynamicBody(short categoryBits, short maskBits, short groupIndex, boolean isSensor){
-        createBody(BodyDef.BodyType.DynamicBody, categoryBits, maskBits, groupIndex, isSensor);
-    }
-    public void createBody(BodyDef.BodyType bodyType){
-        createBody(bodyType, WORLD_BITS, WORLD_BITS, NO_GROUP, false);
-    }
-    public void createBody(BodyDef.BodyType bodyType, short categoryBits, short maskBits, short groupIndex, boolean isSensor){
-        float width = getWidth();
-        float height = getHeight();
-        BodyDef bodyDef = new BodyDef();
-        bodyDef.type = bodyType;
-        // bodyDef.linearDamping = 20f;
-        bodyDef.position.set((x+(width/2))*METRES_PER_PIXEL,
-                             (y+(height/2))*METRES_PER_PIXEL);
-
-        PolygonShape boundingBox = new PolygonShape();
-        boundingBox.setAsBox(((width/2))*METRES_PER_PIXEL,
-                             ((height/2))*METRES_PER_PIXEL);
-
-        FixtureDef fixtureDef = new FixtureDef();
-        fixtureDef.shape = boundingBox;
-        fixtureDef.isSensor = isSensor;
-        
-        fixtureDef.filter.categoryBits = categoryBits;
-        fixtureDef.filter.maskBits = maskBits;
-        fixtureDef.filter.groupIndex = groupIndex;
-
-        body = parent.world.createBody(bodyDef);
-        Fixture fixture = body.createFixture(fixtureDef);
-        fixture.setUserData(this);
-        body.setUserData(this);
-        boundingBox.dispose();
-    }
-    
 
     /**
      * Returns the x coordinate of the entity
      * @return the x coordinate of this Entity
      */
     public float getX() {
-        return getPosition().x;
+        return x;
     }
 
     /**
@@ -117,55 +69,16 @@ public abstract class Entity {
      * @return the y coordinate of this Entity
      */
     public float getY() {
-        return getPosition().y;
+        return y;
     }
     
     public Vector2 getPosition() {
-        return getCentre()
-        .sub(getWidth()/2f, getHeight()/2f);
+        return new Vector2(x, y);
     }
     
     public Vector2 getCentre(){
-        return body.getPosition().scl(PIXELS_PER_METRE);
-    }
-
-    /**
-     * Returns the x velocity of the entity
-     * @return the x velocity of this MobileEntity in pixels per second
-     */
-    public float getVelocityX() {
-        return getVelocity().x;
-    }
-
-    /**
-     * Returns the y velocity of the entity
-     * @return the y coordinate of this MobileEntity in pixels per second
-     */
-    public float getVelocityY() {
-        return getVelocity().y;
-    }
-    
-    public Vector2 getVelocity() {
-        return body.getLinearVelocity().scl(PIXELS_PER_METRE);
-    }
-    
-    public Vector2 getPhysicsVelocity() {
-        return body.getLinearVelocity();
-    }
-    
-    public void setVelocity(Vector2 targetVelocity) {
-        setVelocity(targetVelocity, 0f);
-    }
-    public void setVelocityClamped(Vector2 targetVelocity) {
-        setVelocity(targetVelocity, 4f);
-    }
-    public void setVelocity(Vector2 targetVelocity, float limit) {
-        Vector2 deltaVelocity = targetVelocity.sub(body.getLinearVelocity());
-        if (limit>0){
-            deltaVelocity.clamp(0, limit);
-        }
-        Vector2 impulse = deltaVelocity.scl(body.getMass());
-        body.applyLinearImpulse(impulse, body.getWorldCenter(), true);
+        return getPosition()
+            .add(getWidth()/2f, getHeight()/2f);
     }
 
     /**
@@ -197,9 +110,7 @@ public abstract class Entity {
     public float angleTo(Vector2 dest){
         return vectorTo(dest).angleRad();
     }
-    // public float angleTo(float x, float y) {
-        // return (float)Math.atan2(y - getY(), x - getX());
-    // }
+    
     public float angleTo(float x, float y) {
         return angleTo(new Vector2(x, y));
     }
@@ -252,14 +163,8 @@ public abstract class Entity {
     public boolean isRemoved() {
         return removed;
     }
-
-    /**
-     * Ensures that this MobileEntity stays within the map area.
-     */
-    protected void checkBounds() {
-    }
     
-    public void onCollision(Entity other) {
+    public void dispose(){
     }
     
     /**
