+++ core/src/com/superduckinvaders/game/entity/Obstacle.java
@@ -8,7 +8,7 @@ import com.badlogic.gdx.physics.box2d.*;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 
-public class Obstacle extends Entity {
+public class Obstacle extends PhysicsEntity {
 
     public Obstacle(Round parent, float x, float y, float width, float height) {
         super(parent, x, y);
