+++ core/src/com/superduckinvaders/game/entity/item/Item.java
@@ -3,12 +3,12 @@ package com.superduckinvaders.game.entity.item;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.superduckinvaders.game.Round;
-import com.superduckinvaders.game.entity.Entity;
+import com.superduckinvaders.game.entity.*;
 
 import com.badlogic.gdx.physics.box2d.*;
 import com.badlogic.gdx.math.Vector2;
 
-public class Item extends Entity {
+public class Item extends PhysicsEntity {
 
     /**
      * The texture for this Item.
